Fixed a type issue

+++ lattices.py
@@ -16,7 +16,7 @@ def add_displacement(r, dr=1e-6):
     return r + np.tensor(disp)
 
 def poisson(N, ndim):
-    r = np.rand(N, ndim) - 0.5
+    r = np.rand(N, ndim, dtype=np.double) - 0.5
     return r
 
 def square(Nside=65,centered=True,disp=0):
